# gen/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Winet-S Server
paths:
...
  /auth/refresh:
    post:
      summary: Refresh access token using httpOnly cookie
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Refresh token invalid or expired
...
components:
  schemas:
...
    LoginResponse:
      type: object
      required:
        - access_token
      properties:
        access_token:
          type: string
...